pyproject.toml
[tool.poetry]
name = "postgres-job-queue"
version = "1.0.0"
description = "Example of a job queue with Postgres and Quart"

license = "None"

authors = [
    "Andrei Taranchenko <drey10@gmail.com>"
]
readme = 'README.md'

repository = "https://github.com/papito/postgres-job-queue"
homepage = "https://renegadeotter.com"

keywords = ['postgres', 'jobs', 'queue']

[tool.poetry.dependencies]
python = "^3.11"
toml = "^0.10.2"
siftlog_py = "0.9.5"
python-dotenv = "0.20.0"
quart = "0.18.3"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.2"
Werkzeug = "2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.2"
pytest-asyncio = "^0.20.3"
mypy = "^0.991"
flake8 = "^3.9.2"
isort = "^5.9.1"
black = "22.3.0"
flake8-print = "^5.0.0"
freezegun = "^1.2.2"
types-toml = "^0.1.3"
